[Civic Auth Integration] Refactoring bitgo folder to be inside of services dir
