+++ packages/backend/src/__tests__/unit/transfer.unit.test.js
@@ -1,4 +1,4 @@
-let TransfersManager = require('../../bitgo/TransfersManager');
+let TransfersManager = require('../../services/bitgo/TransfersManager');
 
 test('10% balance calculation using eth wallet', (done) => {
   let tenPercent = TransfersManager._getTenPercentOfBalance(mockEthWallet());
