+++ packages/backend/src/services/bitgo/BitGoSession.js
@@ -1,6 +1,6 @@
 const BitGoJS = require('bitgo');
 
-const { ACCESS_TOKEN } = require('../secureConstants');
+const { ACCESS_TOKEN } = require('../../secureConstants');
 const WalletsManager = require('./WalletsManager');
 const TransfersManager = require('./TransfersManager');
