+++ packages/backend/src/services/bitgo/TransfersManager.js
@@ -1,4 +1,4 @@
-const { RICH_WALLET_ID, WALLET_PASSPHRASE } = require('../secureConstants');
+const { RICH_WALLET_ID, WALLET_PASSPHRASE } = require('../../secureConstants');
 
 class TransfersManager {
